requirements.txt
alembic==1.12.0
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Admin==1.6.1
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-Redmail==0.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.2
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.1
psycopg2-binary==2.9.8
PyJWT==2.8.0
python-dotenv==1.0.0
redmail==0.6.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==2.3.7
WTForms==3.0.1
